ad-joining/register-computer/requirements.txt
flask>=1.0.3
requests>=2.22
ldap3>=2.5.2
google-auth>=1.6.2
google-auth-httplib2>=0.0.3
google-api-python-client>=1.7.8
google-cloud-kms>=1.0.0
oauth2client>=4.1.3
dnspython>=1.16
